refactor(onboarding): move onboarding into it's own component folder
my oh my many things use that "welcomeTemplate" component.
