+++ app/components/onboard/WelcomeTemplate.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react' // eslint-disable-line
 import { remote } from 'electron'
-import Spinner from './chrome/Spinner'
+import Spinner from '../chrome/Spinner'
 import classNames from 'classnames'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faArrowRight } from '@fortawesome/free-solid-svg-icons'
 
-export const logo = require('../assets/qri-blob-logo-small.png') // eslint-disable-line
+export const logo = require('../../assets/qri-blob-logo-small.png') // eslint-disable-line
 
 interface WelcomeTemplateProps {
   onAccept?: () => any
