+++ app/routes.tsx
@@ -2,8 +2,8 @@
 import React from 'react'
 import { Switch, Route, Redirect } from 'react-router-dom'
 import { ipcRenderer } from 'electron'
-import Welcome from './components/Welcome'
-import Signup from './components/Signup'
+import Welcome from './components/onboard/Welcome'
+import Signup from './components/onboard/Signup'
 import Signin from './components/Signin'
 import CollectionContainer from './containers/CollectionContainer'
 import WorkbenchContainer from './containers/WorkbenchContainer'
